+++ test/testdata/faker/openapi3.yml
@@ -1,0 +1,43 @@
+openapi: 3.0.3
+
+info:
+  title: Users dummy API
+  version: 0.1.0
+
+paths:
+  /users/{userId}:
+    get:
+      parameters:
+        - in: path
+          name: userId
+          description: ''
+          required: true
+          schema:
+            type: string
+      responses:
+        '200':
+          description: ''
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/User'
+
+components:
+  schemas:
+    User:
+      type: object
+      required:
+        - id
+        - firstName
+        - lastName
+      properties:
+        id:
+          type: string
+          format: uuid
+          x-faker: uuid
+        firstName:
+          type: string
+          x-faker: firstName
+        lastName:
+          type: string
+          x-faker: lastName
