+++ test/testdata/faker/pact.json
@@ -1,0 +1,36 @@
+{
+  "consumer": {
+    "name": "consumer"
+  },
+  "provider": {
+    "name": "dummy"
+  },
+  "interactions": [
+    {
+      "description": "",
+      "request": {
+        "method": "GET",
+        "path": "/users/example"
+      },
+      "response": {
+        "status": 200,
+        "body": {
+          "firstName": "",
+          "id": "",
+          "lastName": ""
+        },
+        "matchingRules": {
+          "$.body.id": {
+            "match": "type"
+          },
+          "$.body.firstName": {
+            "match": "type"
+          },
+          "$.body.lastName": {
+            "match": "type"
+          }
+        }
+      }
+    }
+  ]
+}
